+++ src/interfaces/Database.ts
@@ -1,0 +1,6 @@
+import type { DatabaseSync } from "node:sqlite";
+
+export default interface Database
+{
+    connectDatabase(): DatabaseSync
+}